Minor bugs fixed. Tested Jacek's charset support but it's still broken.
svn path=/trunk/kdenetwork/kmail/; revision=6811

+++ kalarmtimer.cpp
@@ -38,7 +38,7 @@ static KAlarmTimerId idCounter = 1;
 static char* time2str(struct timeval& tv)
 {
   static char str[32];
-  sprintf (str, "%ld.%ld", tv.tv_sec, tv.tv_usec);
+  sprintf (str, "%ld.%ld", (long)tv.tv_sec, (long)tv.tv_usec);
   return str;
 }
